+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.0
+
+- Added new widget `InputSlider`
+- Added new widget `InputSpinner`
+- Added some reusable validators: `NotNull, Min, MinLength, Max, MaxLength`
+- Validators can have individual error messages with named parameter `message`
+
 ## 0.1.0
 
 - Initial version