+++ example/CHANGELOG.md
@@ -1,3 +1,8 @@
-## 1.0.0
+## 0.2.0
 
-- Initial version, created by Stagehand
+- Added input fields `InputSlider` and `InputSpinner`
+- Added some validator examples
+
+## 0.1.0
+
+- Initial version